api option to deserialize fingerprints separately
the `binary_fuse(8|16)_deserialize_header` functions only deserialize
the core struct fields, but do not `malloc` new memory for the
Fingerprints nor do they set the Fingerprints pointer.

This places memory control in the hands of the user, and facilitates use
cases such as a memory mapped file.

the existing `binary_fuse(8|16)_deserialize` functions are unaffected
but call the new functions internally to avoid code duplication.

resolves https://github.com/FastFilter/xor_singleheader/issues/58