Store the original size argument in struct binary_fuse{8,16}_s
... in preparation of a more compact serialization format: All other
parameters except for the Seed are derived from the size parameter.

The drawback is that this format is sensitive to changes of
binary_fuse8_allocate().

Due to alignment, this does not need any more space on 64bit.
(There were 5 32bit values inbetween two 64bit values)

Yet formally, this is a breaking change of the in-core format, which
should not be used to store information across versions. See follow up
commits for new compact serialization formats.