Add pack and unpack of a more compact serialization format (#68)
* Store the original size argument in struct binary_fuse{8,16}_s

... in preparation of a more compact serialization format: All other
parameters except for the Seed are derived from the size parameter.

The drawback is that this format is sensitive to changes of
binary_fuse8_allocate().

Due to alignment, this does not need any more space on 64bit.
(There were 5 32bit values inbetween two 64bit values)

Yet formally, this is a breaking change of the in-core format, which
should not be used to store information across versions. See follow up
commits for new compact serialization formats.

* Add {xor,binary_fuse}{8,16}_{pack,unpack} serialization formats.

Rationale:

As mentioned in the previous commit, for binary_fuse filters, we do not
need to save values derived from the size, saving 5 x sizeof(uint32_t).

For both filter implementations, we add a bitmap to indicate non-zero
fingerprint values. This adds 1/{8,16} of the fingerprint array size,
but saves one or two bytes for each zero fingerprint.

The net result is a packed format which can not be compressed further by
zlib for the bundled unit tests.

Note that this format is incompatible with the existing _serialize()
format and, in the case of binary_fuse, sensitive to changes of the
derived parameters in _allocate.

Interface:

We add _pack_bytes() to match _serialization_bytes(). _pack() and
_unpack() match _serialize() and _deserialize().

The existing _{de,}serialize() interfaces take a buffer pointer only and
thus implicitly assume that the buffer will be of sufficient size. For
the new functions, we add a size_t parameter indicating the size of the
buffer and check its bounds in the implementation.

_pack returns the used size or zero for "does not fit", so when called
with a buffer of arbitrary size, the used space or error condition can
be determined without an additional call to _pack_bytes(), avoiding
duplicate work.

Implementation:

We add some XOR_bitf_* macros to address words and individual bits of
bitfields.

The XOR_ser and XOR_deser macros have the otherwise repeated code for
bounds checking and the actual serialization.

Because the implementations for the 8 and 16 bit words are equal except
for the data type, we add macros and create the actual functions by
expanding the macros with the possible data types.

Alternatives considered:

Compared to _{de,}serialize(), the new functions need to copy individual
fingerprint words rather than the whole array at once, which is less
efficient. Therefor, an implementation using Duff's Device with
branchless code was attempted but dismissed because avoiding
out-of-bounds access would require an over-allocated buffer.

* Adjust unit tests to new _{un,}pack() interface

To exercise the new code without too much of a change to the existing
unit test, we change the signature of _{un,}serialize_gen() to take an
additional (const) size_t argument, which we ignore for
_{un,}serialize().

We add to the reported metrics absolute and relative size information
for the "in-core" and "wire" format, the latter jointly referencing to
_{un,}serialize() and _{un,}pack().

* Document the new _{un,}pack() interface

* tuning the wording and adding a spaceusage benchmark

* changing the wording.

* rewording.

* explicit casts

---------

Co-authored-by: Daniel Lemire <daniel@lemire.me>